add auth routes

+++ Backend/routes/chat.js
@@ -1,6 +1,7 @@
 import express from "express";
 import Thread from "../models/Thread.js";
 import getOpenAIAPIResponse from "../utils/openai.js";
+import authMiddleware from "../middleware/authMiddleware.js";
 
 const router = express.Router();
 
@@ -21,9 +22,9 @@ router.post("/test", async(req, res) => {
 });
 
 //Get all threads
-router.get("/thread", async(req, res) => {
+router.get("/thread", authMiddleware, async(req, res) => {
     try {
-        const threads = await Thread.find({}).sort({updatedAt: -1});
+        const threads = await Thread.find({ userId: req.userId }).sort({updatedAt: -1});
         //descending order of updatedAt...most recent data on top
         res.json(threads);
     } catch(err) {
@@ -32,11 +33,11 @@ router.get("/thread", async(req, res) => {
     }
 });
 
-router.get("/thread/:threadId", async(req, res) => {
+router.get("/thread/:threadId", authMiddleware, async(req, res) => {
     const {threadId} = req.params;
 
     try {
-        const thread = await Thread.findOne({threadId});
+        const thread = await Thread.findOne({threadId, userId: req.userId});
 
         if(!thread) {
             res.status(404).json({error: "Thread not found"});
@@ -49,11 +50,11 @@ router.get("/thread/:threadId", async(req, res) => {
     }
 });
 
-router.delete("/thread/:threadId", async (req, res) => {
+router.delete("/thread/:threadId", authMiddleware, async (req, res) => {
     const {threadId} = req.params;
 
     try {
-        const deletedThread = await Thread.findOneAndDelete({threadId});
+        const deletedThread = await Thread.findOneAndDelete({threadId, userId: req.userId});
 
         if(!deletedThread) {
             res.status(404).json({error: "Thread not found"});
@@ -67,7 +68,7 @@ router.delete("/thread/:threadId", async (req, res) => {
     }
 });
 
-router.post("/chat", async(req, res) => {
+router.post("/chat", authMiddleware, async(req, res) => {
     const {threadId, message} = req.body;
 
     if(!threadId || !message) {
@@ -75,14 +76,15 @@ router.post("/chat", async(req, res) => {
 }
 
     try {
-        let thread = await Thread.findOne({threadId});
+        let thread = await Thread.findOne({threadId, userId: req.userId});
 
         if(!thread) {
             //create a new thread in Db
             thread = new Thread({
                 threadId,
                 title: message,
-                messages: [{role: "user", content: message}]
+                messages: [{role: "user", content: message}],
+                userId: req.userId
             });
         } else {
             thread.messages.push({role: "user", content: message});
